+++ docs/examples/06_backends_kinematics/files/02_iter_ik_pybullet.py
@@ -1,22 +1,12 @@
 from compas.geometry import Frame
-from compas.robots import LocalPackageMeshLoader
-import compas_fab
 from compas_fab.backends import AnalyticalInverseKinematics
 from compas_fab.backends import PyBulletClient
-
-
-urdf_filename = compas_fab.get('universal_robot/ur_description/urdf/ur5.urdf')
-srdf_filename = compas_fab.get('universal_robot/ur5_moveit_config/config/ur5.srdf')
 
 frame_WCF = Frame((0.381, 0.093, 0.382), (0.371, -0.292, -0.882), (0.113, 0.956, -0.269))
 
 
 with PyBulletClient(connection_type='direct') as client:
-
-    # Load UR5
-    loader = LocalPackageMeshLoader(compas_fab.get('universal_robot'), 'ur_description')
-    robot = client.load_robot(urdf_filename, [loader])
-    client.load_semantics(robot, srdf_filename)
+    robot = client.load_ur5(load_geometry=True)
 
     ik = AnalyticalInverseKinematics(client)
     # set a new IK function