+++ docs/examples/06_backends_kinematics/files/04_cartesian_path_analytic_pybullet.py
@@ -1,11 +1,7 @@
 import matplotlib.pyplot as plt
 from compas.geometry import Frame
-from compas.robots import LocalPackageMeshLoader
-import compas_fab
-from compas_fab.backends import AnalyticalPyBulletClient
 
-urdf_filename = compas_fab.get('universal_robot/ur_description/urdf/ur5.urdf')
-srdf_filename = compas_fab.get('universal_robot/ur5_moveit_config/config/ur5.srdf')
+from compas_fab.backends import AnalyticalPyBulletClient
 
 frames_WCF = [Frame((0.407, 0.073, 0.320), (0.922, 0.000, 0.388), (0.113, 0.956, -0.269)),
               Frame((0.404, 0.057, 0.324), (0.919, 0.000, 0.394), (0.090, 0.974, -0.210)),
@@ -14,9 +10,7 @@ frames_WCF = [Frame((0.407, 0.073, 0.320), (0.922, 0.000, 0.388), (0.113, 0.956,
               Frame((0.376, 0.087, 0.299), (0.850, 0.000, 0.528), (0.184, 0.937, -0.296))]
 
 with AnalyticalPyBulletClient(connection_type='direct') as client:
-    loader = LocalPackageMeshLoader(compas_fab.get('universal_robot'), 'ur_description')
-    robot = client.load_robot(urdf_filename, [loader])
-    client.load_semantics(robot, srdf_filename)
+    robot = client.load_ur5(load_geometry=True)
 
     options = {"solver": "ur5", "check_collision": True}
     start_configuration = list(robot.iter_inverse_kinematics(frames_WCF[0], options=options))[-1]